requirements.txt
torch
torchvision
Pillow
PyYAML
tqdm
scikit-learn
matplotlib
seaborn
numpy
